requirements.txt
Django==2.2.3
django-ckeditor==5.7.1
django-js-asset==1.2.2
Pillow==6.1.0
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
